Created a Main Menu and basic HUD.
Set the default game mode to a menu based game mode (which loads the
main menu). When the player start a new game it loads the demo level
which overrides the game mode and displays the player HUD.

Use the number pad to test the various aspects of the HUD (threat meter,
objectives, and a toast system).